# openapi.yaml
openapi: 3.1.0
info:
  title: Take home tech task
  version: v1.0.0
  description: |-
    We want you to create a REST API for Eagle Bank which conforms to this OpenAPI specification which allows a user to create, fetch, update and delete a bank account and deposit or withdraw money from the account. These will be stored as transactions against a bank account which be retrieved but not modified or deleted.
...
paths:
...
  /v1/auth/token:
    post:
      tags:
        - auth
      description: Authenticate a user and return a JWT bearer token
      operationId: authenticateUser
      requestBody:
        description: Authenticate with user credentials
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthenticateUserRequest'
        required: true
      responses:
        '200':
          description: Authentication succeeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationResponse'
        '400':
          description: Invalid details supplied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '401':
          description: Email or password is invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: An unexpected error occurred
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    AuthenticateUserRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          writeOnly: true
    AuthenticationResponse:
      type: object
      required:
        - token
      properties:
        token:
          type: string
          description: JWT bearer token to pass in the Authorization header
...
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT